+++ lib/msal-angular/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @azure/msal-angular
 
-This log was last generated on Mon, 12 Sep 2022 18:19:32 GMT and should not be manually modified.
+This log was last generated on Mon, 10 Oct 2022 22:27:03 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 2.4.5
+
+Mon, 10 Oct 2022 22:27:03 GMT
+
+### Patches
+
+- Bump @azure/msal-browser to v2.30.0
+
+## 2.4.4
+
+Mon, 03 Oct 2022 22:12:26 GMT
+
+### Patches
+
+- Bump @azure/msal-browser to v2.29.0
 
 ## 2.4.3